requirements.txt
asgiref==3.2.7
Django==2.2.12
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==12.1.0
pytz==2020.1
sqlparse==0.3.1
asgiref==3.2.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==12.1.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
asgiref==3.2.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==12.1.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
asgiref==3.2.7
Django==2.2.12
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==12.1.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
